refactor(layermanager): reorganize lib folders to match monorepo geo convention
Nest the six component dirs under lib/geo/ (mirroring tinacms-portal-monorepo
packages/blocks/src/geo/) and split the styles manifest into a pure aggregator
at lib/styles/components-geo/index.scss + a thin theme manifest at
lib/styles/index.scss. Update cross-domain import paths in components and the
lib barrel. Rename css-modules.d.ts to scss-imports.d.ts to reflect its
actual purpose (ambient declaration for side-effect .scss imports).

Lib remains self-contained: lib/styles/theme-tokens.scss still owns the
USWDS configuration for this lib's compiled CSS. The MMGIS-global theme
lift is a separate concern on a follow-up branch.
