+++ src/essence/Tools/LayerManager/lib/geo/CategoricalGraphic/CategoricalGraphic.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useState, useCallback, useRef, type MouseEvent } from 'react'
-import type { CategoricalStop } from '../types'
+import type { CategoricalStop } from '../../types'
 
 export type CategoricalGraphicProps = {
     stops: CategoricalStop[]
