+++ src/essence/Tools/LayerManager/lib/index.ts
@@ -1,9 +1,9 @@
-export { LayerManagerPanel, type LayerManagerPanelProps } from './LayerManagerPanel/LayerManagerPanel'
-export { LayerLegendList, type LayerLegendListProps } from './LayerLegendList/LayerLegendList'
-export { LayerLegend, type LayerLegendProps } from './LayerLegend/LayerLegend'
-export { GradientGraphic, type GradientGraphicProps } from './GradientGraphic/GradientGraphic'
-export { CategoricalGraphic, type CategoricalGraphicProps } from './CategoricalGraphic/CategoricalGraphic'
-export { ColormapControl, type ColormapControlProps } from './ColormapControl/ColormapControl'
+export { LayerManagerPanel, type LayerManagerPanelProps } from './geo/LayerManagerPanel/LayerManagerPanel'
+export { LayerLegendList, type LayerLegendListProps } from './geo/LayerLegendList/LayerLegendList'
+export { LayerLegend, type LayerLegendProps } from './geo/LayerLegend/LayerLegend'
+export { GradientGraphic, type GradientGraphicProps } from './geo/GradientGraphic/GradientGraphic'
+export { CategoricalGraphic, type CategoricalGraphicProps } from './geo/CategoricalGraphic/CategoricalGraphic'
+export { ColormapControl, type ColormapControlProps } from './geo/ColormapControl/ColormapControl'
 
 export type { Layer, LegendType, CategoricalStop, CogData, Colormap } from './types'
