+++ src/essence/Tools/LayerManager/lib/styles/components-geo/categorical-graphic.scss
