+++ src/essence/Tools/LayerManager/lib/styles/components-geo/colormap-control.scss
