+++ src/essence/Tools/LayerManager/lib/styles/components-geo/gradient-graphic.scss
