+++ src/essence/Tools/LayerManager/lib/styles/components-geo/index.scss
@@ -1,0 +1,10 @@
+// Aggregator for geo component partials. Matches the monorepo convention at
+// tinacms-portal-monorepo:packages/blocks/src/styles/components-geo/index.scss
+// — pure listing, no theme-loading. The theme manifest at ../index.scss must
+// configure theme-tokens and load USWDS before forwarding this aggregator.
+@use 'layer-manager-panel';
+@use 'layer-legend-list';
+@use 'layer-legend';
+@use 'gradient-graphic';
+@use 'categorical-graphic';
+@use 'colormap-control';
