+++ src/essence/Tools/LayerManager/lib/styles/components-geo/layer-legend-list.scss
