+++ src/essence/Tools/LayerManager/lib/styles/components-geo/layer-manager-panel.scss
