+++ src/essence/Tools/LayerManager/lib/styles/index.scss
@@ -1,8 +1,7 @@
+// LayerManager lib theme manifest. Self-contained: configures USWDS via the
+// local theme-tokens, loads USWDS framework CSS, and pulls in component
+// partials via the geo aggregator. Mirrors the monorepo's per-theme manifest
+// shape at tinacms-portal-monorepo:packages/blocks/src/styles/disasters/index.scss.
 @forward 'theme-tokens';
 @forward 'uswds';
-@forward 'components/layer-manager-panel';
-@forward 'components/layer-legend-list';
-@forward 'components/layer-legend';
-@forward 'components/gradient-graphic';
-@forward 'components/categorical-graphic';
-@forward 'components/colormap-control';
+@forward 'components-geo';
